+++ CHANGELOG.md
@@ -5,8 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- `useContactReaderIdentificationFilter` marked as deprecated.
 ### Fixed
 - Default reader type identification filters.
+- Logging of the available readers names in the `AbstractPcscPluginAdapter` class.
 
 ## [2.1.0] - 2022-07-25
 ### Added